requirements.txt
streamlit
google-generativeai
pandas
pymupdf
plotly
faiss-cpu
sentence-transformers